requirements.txt
boto3
PyGithub
google-auth
azure-identity
